pyproject.toml
[tool.poetry]
name = "devops_agent"
version = "0.1.0"
description = "Devops GPT Devops Agent"
authors = ["Zelarsoft Pvt. Ltd. <https://github.com/zelarhq>"]
license = "MIT"

[tool.poetry.group.dev.dependencies]
pyclean = "^3.1.0"
black = "^25.1.0"
isort = "^6.0.1"

[project]
name = "devops_agent"
version = "0.0.1"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
langchain-openai = "0.2.3"
langchain-community = "^0.3.1"
langchain-anthropic = "0.2.3"
langchain-google-genai = "2.0.5"
langchain = "0.3.4"
openai = "^1.52.1"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"
requests = "^2.32.3"
html2text = "^2024.2.26"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
pyinstaller = "^6.11.1"
pydantic = "*"
loguru = "^0.7.3"
langgraph-checkpoint = "^2.0.10"
langgraph-checkpoint-postgres = "^2.0.13"
langchain-postgres = "^0.0.12"
copilotkit = "^0.1.34"
psycopg-binary = "^3.2.4"
langchain-text-splitters = "^0.3.5"
beautifulsoup4 = "^4.12.3"
atlassian-python-api = "^3.41.19"
lxml = "^5.3.1"
unstructured = "^0.16.23"
markdown = "^3.7"
psycopg2-binary = "^2.9.10"

[tool.poetry.scripts]
devops_agent = "app:main"
